test: reorganization. update `make unit-test` to `make test` (#1267)
