+++ .github/workflows/codecov.yaml
@@ -19,8 +19,8 @@ jobs:
           go-version-file: './go.mod'
           cache-dependency-path: './go.sum'
 
-      - name: Unit Tests
-        run: make unit-test
+      - name: Tests
+        run: make test
 
       - name: upload coverage to codecov
         uses: codecov/codecov-action@eaaf4bedf32dbdc6b720b63067d99c4d77d6047d # v3.1.4
